[FEATURE] Create function fetchRecommendMusics() and unit test

+++ app/src/main/java/com/moi/lime/vo/MusicUrlEntity.kt
@@ -3,15 +3,15 @@ package com.moi.lime.vo
 import com.google.gson.annotations.SerializedName
 
 data class MusicUrlEntity(@SerializedName("code")
-                          val code: Int = 0,
+                          val code: Long = 0,
                           @SerializedName("expi")
-                          val expi: Int = 0,
+                          val expi: Long = 0,
                           @SerializedName("flag")
-                          val flag: Int = 0,
+                          val flag: Long = 0,
                           @SerializedName("level")
                           val level: String = "",
                           @SerializedName("fee")
-                          val fee: Int = 0,
+                          val fee: Long = 0,
                           @SerializedName("type")
                           val type: String = "",
                           @SerializedName("canExtend")
@@ -21,18 +21,18 @@ data class MusicUrlEntity(@SerializedName("code")
                           @SerializedName("gain")
                           val gain: Double = 0.0,
                           @SerializedName("br")
-                          val br: Int = 0,
+                          val br: Long = 0,
                           @SerializedName("uf")
                           val uf: String? = null,
                           @SerializedName("size")
-                          val size: Int = 0,
+                          val size: Long = 0,
                           @SerializedName("encodeType")
                           val encodeType: String = "",
                           @SerializedName("id")
-                          val id: Int = 0,
+                          val id: Long = 0,
                           @SerializedName("md5")
                           val md: String = "",
                           @SerializedName("payed")
-                          val payed: Int = 0,
+                          val payed: Long = 0,
                           @SerializedName("freeTrialInfo")
                           val freeTrialInfo: String? = null)
